Refactor CSS and components, and addition of the Column system
