+++ src/components/login/LaunchPadBanner.scss
@@ -1,0 +1,9 @@
+.banner {
+    position: absolute;
+    left: 20px;
+    top: 20px;
+    .icon {
+        border-radius: 5px;
+        background-color: white;
+    }
+}
